+++ NEWS.md
@@ -1,5 +1,7 @@
 # purrr (development version)
 
+* `map_vec()` no longer fails on empty named lists (#1206).
+
 * Formatted strings for the progress bar could only access the global environment. `map()`, `map2()`, and `pmap()` have been updated to add  `caller = .purrr_user_env` to `.progress` by default, allowing formatted strings to access to the current/parent environment (@jcolt45, #1078).
 
 * `as_mapper.default()` optimized by removing special named argument handling for primitive functions (@mtcarsalot, #1088).